+++ CHANGELOG.md
@@ -15,6 +15,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ### Fixed
 * Small issues in CIFAR Example
 * Small `imgaug` issue in `examples/sseg_camvid.py` and `examples/segmentation.py`
+* FitHarn no longer fails when loaders are missing batch sizes
 
 
 ## Version 0.5.1